Work around YResolution rounding issue that can cause more pixels than desired to be imaged.
Shan explained: "The YResolution is calculated based on the Y dimension (in µm) and pixel size (in nm). However, it is rounded to multiples of 4 during acquisition due to some issues to synchronize the high speed NI cards."

The FileLength header value includes bytes from the rounded-up YResolution region but the dat readers ignore those bytes because they only load (XResolution * YResolution * ChanNum * bytes-per-pixel) bytes as image data.  As an example, a 9125x8250 dat actually has image data for a 9125x8252 area since 8250 is not evenly divisible by 4.  The 9125x8250 dat FileLength header value will include the 2 extra pixel rows, but the ndarray of pixel data from the reader will not include those 2 extra rows.

To support round-trip conversions of dat to h5 and back to dat, we now ignore the idea of a recipe and instead think of the data as having a header, pixels, and a footer where the footer contains the recipe and possibly a few extra rows of pixels when acquisition rounding occurs.